Break out referrer URL helper tests
